deps!: upgrade Magick.NET-Q8-x64 7.5.0.1 → 14.11.1
Resolves 80+ security advisories (heap overflows, DoS, buffer overreads,
code injection) and eliminates NETSDK1206 win7-x64 RID warning.

API migration changes:
- Drawables class moved from ImageMagick to ImageMagick.Drawing namespace
- ISkinModule.GetDrawables return type: Drawables → IDrawables<byte>
- StrokeAntialias(bool)/TextAntialias(bool) → Enable/Disable variants
- MagickImage constructor: int → uint for dimensions
- GetPixels().ToColor() returns IMagickColor<byte>? instead of MagickColor

15 files changed across Skin/ and ProfileCardService

+++ DiscordBot/Services/ProfileCardService.cs
@@ -116,7 +116,7 @@ public class ProfileCardService
                     profile.Picture = new MagickImage($"{_settings.AssetsRootPath}/images/default.png");
                 }
 
-            profile.Picture.Resize(skin.AvatarSize, skin.AvatarSize);
+            profile.Picture.Resize((uint)skin.AvatarSize, (uint)skin.AvatarSize);
             profileCard.Add(background);
 
             foreach (var layer in skin.Layers)
@@ -130,7 +130,7 @@ public class ProfileCardService
                     background.Composite(image, (int)layer.StartX, (int)layer.StartY, CompositeOperator.Over);
                 }
 
-                var l = new MagickImage(MagickColors.Transparent, (int)layer.Width, (int)layer.Height);
+                var l = new MagickImage(MagickColors.Transparent, (uint)layer.Width, (uint)layer.Height);
                 foreach (var module in layer.Modules) module.GetDrawables(profile).Draw(l);
 
                 background.Composite(l, (int)layer.StartX, (int)layer.StartY, CompositeOperator.Over);
